Moved business rules from Controllers to Services

+++ src/main/java/org/springframework/samples/petclinic/service/ReviewService.java
@@ -6,8 +6,17 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataAccessException;
 import org.springframework.data.jpa.repository.Modifying;
+import org.springframework.samples.petclinic.model.Authorities;
 import org.springframework.samples.petclinic.model.Review;
+import org.springframework.samples.petclinic.model.User;
 import org.springframework.samples.petclinic.repository.ReviewRepository;
+import org.springframework.samples.petclinic.service.exceptions.CantDeleteReviewException;
+import org.springframework.samples.petclinic.service.exceptions.CantEditReviewException;
+import org.springframework.samples.petclinic.service.exceptions.CantWriteReviewException;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -17,38 +26,68 @@ public class ReviewService {
 	@Autowired
 	private ReviewRepository reviewRepo;
 
+	@Autowired
+	private ReadBookService readBookService;
+
+	@Autowired
+	private UserService userService;
+
+	@Autowired
+	private AuthoritiesService authoritiesService;
+
 	@Transactional
 	public List<Integer> getReviewsIdFromBook(int bookId) throws DataAccessException {
 		return this.reviewRepo.getReviewsIdFromBook(bookId);
 	}
-	
+
 	@Transactional
 	public List<Review> getReviewsFromBook(int bookId) throws DataAccessException {
 		return this.reviewRepo.getReviewsFromBook(bookId);
 	}
-	
+
 	@Transactional
 	@Modifying
-	public void deleteReviewById(int reviewId) throws DataAccessException {
-		this.reviewRepo.deleteReviewById(reviewId);
+	public void deleteReviewById(int reviewId, String username) throws DataAccessException, CantDeleteReviewException {
+		Boolean canDeleteReview = this.canDeleteReview(reviewId, username);
+		if(canDeleteReview) {
+			this.reviewRepo.deleteReviewById(reviewId);
+		}else {
+			throw new CantDeleteReviewException();
+		}
 	}
-	
+
 	@Transactional
 	public Review findReviewById(int reviewId) throws DataAccessException {
 		return this.reviewRepo.findById(reviewId);
 	}
-	
+
 	@Transactional 
-	public Boolean existsReviewById(int reviewId) throws DataAccessException{
+	public Boolean existsReviewById(int reviewId) throws DataAccessException {
 		return this.reviewRepo.existsById(reviewId);
 	}
-	
+
 	@Transactional
 	@Modifying
-	public void save(Review review) throws DataAccessException{
-		this.reviewRepo.save(review);
+	public void writeReview(Review review, String username) throws DataAccessException, CantWriteReviewException {
+		Boolean canWriteReview = this.canWriteReview(review.getBook().getId(), username);
+		if(canWriteReview) {
+			this.reviewRepo.save(review);
+		}else {
+			throw new CantWriteReviewException();
+		}
 	}
-	
+
+	@Transactional
+	@Modifying
+	public void editReview(Review review, String username) throws DataAccessException, CantEditReviewException {
+		Boolean isMine = this.reviewIsMine(review.getId(), username);
+		if(isMine) {
+			this.reviewRepo.save(review);
+		}else {
+			throw new CantEditReviewException();
+		}	
+	}
+
 	@Transactional
 	public Boolean alreadyReviewedBook(int bookId, String username) throws DataAccessException{
 		Review review = this.reviewRepo.getReviewByBookIdAndUsername(bookId, username);
@@ -58,9 +97,9 @@ public class ReviewService {
 			return true;
 		}
 	}
-	
+
 	@Transactional
-	public Boolean reviewIsMine(int reviewId, String username) throws DataAccessException{
+	public Boolean reviewIsMine(int reviewId, String username) throws DataAccessException {
 		Review review = this.reviewRepo.findById(reviewId);
 		if(review.getUser().getUsername().equals(username)) {
 			return true;
@@ -68,6 +107,34 @@ public class ReviewService {
 			return false;
 		}
 	}
-	
+
+	public Boolean canWriteReview(int bookId, String username) throws DataAccessException {
+		Boolean bookIsRead = this.readBookService.esReadBook(bookId, username);
+		Boolean alreadyReviewed = this.alreadyReviewedBook(bookId, username);
+		if(bookIsRead && !alreadyReviewed) {
+			return true;
+		}else {
+			return false;
+		}
+	}
+
+	public Boolean canDeleteReview(int reviewId, String username) {
+		//Solo las borra el admin o el duenyo
+		Boolean isMine = this.reviewIsMine(reviewId, username);
+
+		List<Authorities> authorities = this.authoritiesService.getAuthoritiesByUsername(username);
+		Boolean imAdmin = false;
+		for(Authorities a: authorities){
+			if(a.getAuthority().equals("admin")){
+				imAdmin = true;
+			}
+		}
+		if(isMine || imAdmin) {
+			return true;
+		}else {
+			return false;
+		}
+
+	}
 
 }
